feat: new command that calls REST and Tooling APIs to get Query Plan data and print it into Output Tab
