+++ packages/salesforcedx-vscode-soql/package.json
@@ -353,6 +353,10 @@
       {
         "command": "sf.data.query.selection",
         "title": "%data_query_selection_text%"
+      },
+      {
+        "command": "sf.data.query.explain.selection",
+        "title": "%query_plan_selection_text%"
       }
     ],
     "menus": {
@@ -379,6 +383,10 @@
         },
         {
           "command": "sf.data.query.selection",
+          "when": "sf:project_opened && editorHasSelection && sf:has_target_org"
+        },
+        {
+          "command": "sf.data.query.explain.selection",
           "when": "sf:project_opened && editorHasSelection && sf:has_target_org"
         }
       ],
