+++ packages/salesforcedx-vscode-soql/package.nls.json
@@ -24,5 +24,6 @@
   "soql_walkthrough_step5_description": "After running a query, you can export the results by clicking the **Save** button in the results viewer. Choose between CSV or JSON format for further analysis. You can also save your SOQL queries as `.soql` files to reuse them later.\n\n💡 **Pro tip**: CSV format is great for spreadsheet analysis, while JSON format preserves data relationships for programmatic use.",
   "soql_walkthrough_step5_alt": "Saving and exporting SOQL query results",
   "data_query_input_text": "SFDX: Execute SOQL Query...",
-  "data_query_selection_text": "SFDX: Execute SOQL Query with Currently Selected Text"
+  "data_query_selection_text": "SFDX: Execute SOQL Query with Currently Selected Text",
+  "query_plan_selection_text": "SFDX: Get SOQL Query Plan with Currently Selected Text"
 }
