+++ packages/salesforcedx-vscode-soql/src/commands/dataQuery.ts
@@ -6,15 +6,7 @@
  */
 import type { Connection } from '@salesforce/core';
 import { getServicesApi, sfProjectPreconditionChecker } from '@salesforce/effect-ext-utils';
-import {
-  CancelResponse,
-  Column,
-  ContinueResponse,
-  createTable,
-  ParametersGatherer,
-  Row,
-  SfCommandlet
-} from '@salesforce/salesforcedx-utils-vscode';
+import { Column, ContinueResponse, createTable, Row, SfCommandlet } from '@salesforce/salesforcedx-utils-vscode';
 import * as Effect from 'effect/Effect';
 import * as vscode from 'vscode';
 import { Utils } from 'vscode-uri';
@@ -22,6 +14,7 @@ import { nls } from '../messages';
 import { channelService } from '../services/channel';
 import { getSoqlRuntime } from '../services/extensionProvider';
 import { getConnection } from '../services/org';
+import { formatErrorMessage, GetQueryAndApiInputs, QueryAndApiInputs } from './queryUtils';
 
 type QueryResult = Awaited<ReturnType<Connection['query']>>;
 
@@ -78,61 +71,6 @@ class DataQueryExecutor {
       });
   }
 }
-
-class GetQueryAndApiInputs implements ParametersGatherer<QueryAndApiInputs> {
-  public async gather(): Promise<CancelResponse | ContinueResponse<QueryAndApiInputs>> {
-    const editor = vscode.window.activeTextEditor;
-
-    let query;
-
-    if (!editor) {
-      const userInputOptions: vscode.InputBoxOptions = {
-        prompt: nls.localize('parameter_gatherer_enter_soql_query')
-      };
-      query = await vscode.window.showInputBox(userInputOptions);
-    } else {
-      const document = editor.document;
-      if (editor.selection.isEmpty) {
-        const userInputOptions: vscode.InputBoxOptions = {
-          prompt: nls.localize('parameter_gatherer_enter_soql_query')
-        };
-        query = await vscode.window.showInputBox(userInputOptions);
-      } else {
-        query = document.getText(editor.selection);
-      }
-    }
-    if (!query) {
-      return { type: 'CANCEL' };
-    }
-
-    query = query
-      .replace('[', '')
-      .replace(']', '')
-      .replaceAll(/(\r\n|\n)/g, ' ');
-
-    const restApi = {
-      api: 'REST' as const,
-      label: nls.localize('REST_API'),
-      description: nls.localize('REST_API_description')
-    };
-
-    const toolingApi = {
-      api: 'TOOLING' as const,
-      label: nls.localize('tooling_API'),
-      description: nls.localize('tooling_API_description')
-    };
-
-    const apiItems = [restApi, toolingApi];
-    const selection = await vscode.window.showQuickPick(apiItems);
-
-    return selection ? { type: 'CONTINUE', data: { query, api: selection.api } } : { type: 'CANCEL' };
-  }
-}
-
-type QueryAndApiInputs = {
-  query: string;
-  api: 'REST' | 'TOOLING';
-};
 
 export const dataQuery = Effect.fn('sf.data.query')(function* () {
   const commandlet = new SfCommandlet(sfProjectPreconditionChecker, new GetQueryAndApiInputs(), new DataQueryExecutor());
@@ -401,60 +339,6 @@ export const convertQueryResultToCSV = (queryResult: QueryResult): string => {
   }
 
   return convertToCSV(queryResult.records);
-};
-
-/** Formats error messages for better user experience */
-export const formatErrorMessage = (error: unknown): string => {
-  // Handle different error formats
-  let errorString: string;
-  if (error instanceof Error) {
-    errorString = error.message;
-  } else if (error && typeof error === 'object' && 'message' in error) {
-    errorString = String(error.message);
-  } else {
-    errorString = String(error);
-  }
-
-  // Check for common error patterns and provide better messages
-  if (errorString.includes('HTTP response contains html content')) {
-    return nls.localize('data_query_error_org_expired');
-  }
-
-  if (errorString.includes('INVALID_SESSION_ID')) {
-    return nls.localize('data_query_error_session_expired');
-  }
-
-  if (errorString.includes('INVALID_LOGIN')) {
-    return nls.localize('data_query_error_invalid_login');
-  }
-
-  if (errorString.includes('INSUFFICIENT_ACCESS')) {
-    return nls.localize('data_query_error_insufficient_access');
-  }
-
-  if (errorString.includes('MALFORMED_QUERY')) {
-    return nls.localize('data_query_error_malformed_query');
-  }
-
-  if (errorString.includes('INVALID_FIELD')) {
-    return nls.localize('data_query_error_invalid_field');
-  }
-
-  if (errorString.includes('INVALID_TYPE')) {
-    return nls.localize('data_query_error_invalid_type');
-  }
-
-  if (errorString.includes('connection') || errorString.includes('network')) {
-    return nls.localize('data_query_error_connection');
-  }
-
-  // For tooling API specific errors
-  if (errorString.includes('tooling') && errorString.includes('not found')) {
-    return nls.localize('data_query_error_tooling_not_found');
-  }
-
-  // Default error message
-  return nls.localize('data_query_error_message', errorString);
 };
 
 /**
